feat(irfan-landing-page): add property hide in listing interface & rename dropdown component
